skill passive reorganization, tooltip improvements
