+++ Content/SkillPassives/Generic/Efficiency.cs
@@ -1,5 +1,4 @@
-﻿using Humanizer;
-using PathOfTerraria.Common.Mechanics;
+﻿using PathOfTerraria.Common.Mechanics;
 using PathOfTerraria.Common.Systems.Skills;
 using PathOfTerraria.Common.Utilities;
 
@@ -8,7 +7,7 @@ namespace PathOfTerraria.Content.SkillPassives.Generic;
 internal class Efficiency : SkillPassive
 {
 	public const float Chance = 0.02f;
-	public override string DisplayTooltip => base.DisplayTooltip.FormatWith(MathUtils.Percent(Chance));
+	public override object[] TooltipArguments => [MathUtils.Percent(Chance)];
 
 	public Efficiency(SkillTree tree) : base(tree)
 	{
