+++ Content/SkillPassives/Generic/FlashFire.cs
@@ -1,5 +1,4 @@
-﻿using Humanizer;
-using PathOfTerraria.Common.Mechanics;
+﻿using PathOfTerraria.Common.Mechanics;
 using PathOfTerraria.Common.Systems.Skills;
 using PathOfTerraria.Common.Utilities;
 
@@ -8,7 +7,7 @@ namespace PathOfTerraria.Content.SkillPassives.Generic;
 internal class FlashFire(SkillTree tree) : SkillPassive(tree)
 {
 	public const float DamageReduction = 0.35f;
-	public override string DisplayTooltip => base.DisplayTooltip.FormatWith(MathUtils.Percent(DamageReduction));
+	public override object[] TooltipArguments => [MathUtils.Percent(DamageReduction)];
 
 	public override void PassiveEffects(ref SkillBuff buff)
 	{
