--- Content/SkillPassives/NovaTree/ThunderClaps.cs
@@ -1,6 +1,0 @@
-﻿using PathOfTerraria.Common.Mechanics;
-using PathOfTerraria.Common.Systems.Skills;
-
-namespace PathOfTerraria.Content.SkillPassives.NovaTree;
-
-internal class ThunderClaps(SkillTree tree) : SkillPassive(tree);
