--- Content/SkillPassives/RainOfArrowsPassives/ColdBlast.cs
@@ -1,6 +1,0 @@
-﻿using PathOfTerraria.Common.Mechanics;
-using PathOfTerraria.Common.Systems.Skills;
-
-namespace PathOfTerraria.Content.SkillPassives.RainOfArrowsPassives;
-
-internal class ColdBlast(SkillTree tree) : SkillPassive(tree);
