--- Content/SkillPassives/RainOfArrowsPassives/ConcussiveBurst.cs
@@ -1,6 +1,0 @@
-﻿using PathOfTerraria.Common.Mechanics;
-using PathOfTerraria.Common.Systems.Skills;
-
-namespace PathOfTerraria.Content.SkillPassives.RainOfArrowsPassives;
-
-internal class ConcussiveBurst(SkillTree tree) : SkillPassive(tree);
