--- Content/SkillPassives/RainOfArrowsPassives/PowerfulSmog.cs
@@ -1,6 +1,0 @@
-﻿using PathOfTerraria.Common.Mechanics;
-using PathOfTerraria.Common.Systems.Skills;
-
-namespace PathOfTerraria.Content.SkillPassives.RainOfArrowsPassives;
-
-internal class PowerfulSmog(SkillTree tree) : SkillPassive(tree);
