--- Content/SkillPassives/RainOfArrowsPassives/SharpenedTips.cs
@@ -1,6 +1,0 @@
-﻿using PathOfTerraria.Common.Mechanics;
-using PathOfTerraria.Common.Systems.Skills;
-
-namespace PathOfTerraria.Content.SkillPassives.RainOfArrowsPassives;
-
-internal class SharpenedTips(SkillTree tree) : SkillPassive(tree);
