--- Content/SkillPassives/RainOfArrowsPassives/TargetLock.cs
@@ -1,6 +1,0 @@
-﻿using PathOfTerraria.Common.Mechanics;
-using PathOfTerraria.Common.Systems.Skills;
-
-namespace PathOfTerraria.Content.SkillPassives.RainOfArrowsPassives;
-
-internal class TargetLock(SkillTree tree) : SkillPassive(tree);
