+++ Localization/fr-FR/Mods.PathOfTerraria.SkillPassives.hjson
@@ -208,22 +208,22 @@ Accelerant: {
 
 BiggerBrood: {
 	// Name: Bigger Brood
-	// Tooltip: +1 egg per level
+	// Tooltip: +{0} egg per level
 }
 
 Gestation: {
 	// Name: Gestation
-	// Tooltip: 10% chance to spawn an additional egg when a locust kills an enemy
+	// Tooltip: "{0}% chance to spawn an additional egg when a locust kills an enemy"
 }
 
 QuickerHatching: {
 	// Name: Quicker Hatching
-	// Tooltip: Eggs spawn 20% quicker per level
+	// Tooltip: Eggs spawn {0}% quicker per level
 }
 
 CarnivorousLarvae: {
 	// Name: Carnivorous Larvae
-	// Tooltip: Additional 2% chance for an egg to spawn in a fallen enemy per level
+	// Tooltip: Additional {0}% chance for an egg to spawn in a fallen enemy per level
 }
 
 Eggsplosion: {
@@ -243,22 +243,22 @@ ViciousBites: {
 
 VolatileInsects: {
 	// Name: Volatile Insects
-	// Tooltip: Antlions explode after hitting their target, dealing 300% fire damage
+	// Tooltip: Antlions explode after hitting their target, dealing {0}% fire damage
 }
 
 InfectedDetonation: {
 	// Name: Infected Detonation
-	// Tooltip: Enemies killed by the {$VolatileInsects.Name} explosion have a 10% chance to explode, dealing 10% of their max health as damage
+	// Tooltip: Enemies killed by the {$VolatileInsects.Name} explosion have a {0}% chance to explode, dealing {1}% of their max health as damage
 }
 
 HeartierExplosions: {
 	// Name: Heartier Explosions
-	// Tooltip: +2% max health converted to {$InfectedDetonation.Name} damage per level
+	// Tooltip: +{0}% max health converted to {$InfectedDetonation.Name} damage per level
 }
 
 CombustableGuts: {
 	// Name: Combustable Guts
-	// Tooltip: +2% chance for {$InfectedDetonation.Name} explosions to occur on enemy death
+	// Tooltip: +{0}% chance for {$InfectedDetonation.Name} explosions to occur on enemy death
 }
 
 OverheatingBugs: {
@@ -268,22 +268,22 @@ OverheatingBugs: {
 
 SuperheatedBugs: {
 	// Name: Superheated Bugs
-	// Tooltip: Bugs and their explosions deal 15% more On Fire! damage
+	// Tooltip: Bugs and their explosions deal {0}% more On Fire! damage
 }
 
 ColdBlooded: {
 	// Name: Cold Blooded
-	// Tooltip: Antlions gain an additional second of life per kill
+	// Tooltip: Antlions gain {0} additional {^0:second;seconds} of life per kill
 }
 
 ThermalConversion: {
 	// Name: Thermal Conversion
-	// Tooltip: +0.5 seconds life gained per kill per level
+	// Tooltip: +{0} seconds life gained per kill per level
 }
 
 FrostbiteMandibles: {
 	// Name: Frostbite Mandibles
-	// Tooltip: Increased cold damage and a 10% chance to inflict Chilled on hit
+	// Tooltip: Increased cold damage and a {0}% chance to inflict Chilled on hit
 }
 
 AggressiveChill: {
