+++ Localization/pt-BR/Mods.PathOfTerraria.SkillPassives.hjson
@@ -20,7 +20,7 @@ Efficiency: {
 
 ThunderClaps: {
 	// Name: Thunder Claps
-	// Tooltip: 5% chance per level to spawn a tiny nova around a hit target
+	// Tooltip: "{0}% chance per level to spawn a tiny nova around a hit target"
 }
 
 ConcurrentBlasts: {
@@ -68,7 +68,7 @@ MoldColony: {
 
 FungalSpread: {
 	// Name: Fungal Spread
-	// Tooltip: Spore explosions are 10% larger per stack
+	// Tooltip: Spore explosions are {0}% larger per stack
 }
 
 LingeringPoison: {
@@ -78,7 +78,7 @@ LingeringPoison: {
 
 PowerfulSmog: {
 	// Name: Powerful Smog
-	// Tooltip: Poison clouds linger a second longer per stack
+	// Tooltip: Poison clouds linger {0} {^0:second;seconds} longer per stack
 }
 
 Megatoxin: {
@@ -88,12 +88,12 @@ Megatoxin: {
 
 ColdBlast: {
 	// Name: Cold Blast
-	// Tooltip: All Explosive Volley explosions inflict 2 seconds of Chilled and deals 100% cold damage
+	// Tooltip: All Explosive Volley explosions inflict {0} seconds of Chilled and deals {1}% cold damage
 }
 
 ShatteringArrows: {
 	// Name: Shattering Arrows
-	// Tooltip: Volleys have twice the amount of projectiles, each dealing 70% of original damage
+	// Tooltip: Volleys have twice the amount of projectiles, each dealing {0}% of original damage
 }
 
 ConcussiveBurst: {
@@ -103,7 +103,7 @@ ConcussiveBurst: {
 
 SlicingShrapnel: {
 	// Name: Slicing Shrapnel
-	// Tooltip: Shatter projectiles all inflict bleeding for 3 seconds
+	// Tooltip: Shatter projectiles all inflict bleeding for {0} seconds
 }
 
 CreepingVines: {
@@ -118,7 +118,7 @@ Quickload: {
 
 SharpenedTips: {
 	// Name: Sharpened Tips
-	// Tooltip: 10% increased armor piercing per stack.
+	// Tooltip: "{0}% increased armor piercing per stack."
 }
 
 TargetLock: {
@@ -208,22 +208,22 @@ Accelerant: {
 
 BiggerBrood: {
 	// Name: Bigger Brood
-	// Tooltip: +1 egg per level
+	// Tooltip: +{0} egg per level
 }
 
 Gestation: {
 	// Name: Gestation
-	// Tooltip: 10% chance to spawn an additional egg when a locust kills an enemy
+	// Tooltip: "{0}% chance to spawn an additional egg when a locust kills an enemy"
 }
 
 QuickerHatching: {
 	// Name: Quicker Hatching
-	// Tooltip: Eggs spawn 20% quicker per level
+	// Tooltip: Eggs spawn {0}% quicker per level
 }
 
 CarnivorousLarvae: {
 	// Name: Carnivorous Larvae
-	// Tooltip: Additional 2% chance for an egg to spawn in a fallen enemy per level
+	// Tooltip: Additional {0}% chance for an egg to spawn in a fallen enemy per level
 }
 
 Eggsplosion: {
@@ -243,22 +243,22 @@ ViciousBites: {
 
 VolatileInsects: {
 	// Name: Volatile Insects
-	// Tooltip: Antlions explode after hitting their target, dealing 300% fire damage
+	// Tooltip: Antlions explode after hitting their target, dealing {0}% fire damage
 }
 
 InfectedDetonation: {
 	// Name: Infected Detonation
-	// Tooltip: Enemies killed by the {$VolatileInsects.Name} explosion have a 10% chance to explode, dealing 10% of their max health as damage
+	// Tooltip: Enemies killed by the {$VolatileInsects.Name} explosion have a {0}% chance to explode, dealing {1}% of their max health as damage
 }
 
 HeartierExplosions: {
 	// Name: Heartier Explosions
-	// Tooltip: +2% max health converted to {$InfectedDetonation.Name} damage per level
+	// Tooltip: +{0}% max health converted to {$InfectedDetonation.Name} damage per level
 }
 
 CombustableGuts: {
 	// Name: Combustable Guts
-	// Tooltip: +2% chance for {$InfectedDetonation.Name} explosions to occur on enemy death
+	// Tooltip: +{0}% chance for {$InfectedDetonation.Name} explosions to occur on enemy death
 }
 
 OverheatingBugs: {
@@ -268,22 +268,22 @@ OverheatingBugs: {
 
 SuperheatedBugs: {
 	// Name: Superheated Bugs
-	// Tooltip: Bugs and their explosions deal 15% more On Fire! damage
+	// Tooltip: Bugs and their explosions deal {0}% more On Fire! damage
 }
 
 ColdBlooded: {
 	// Name: Cold Blooded
-	// Tooltip: Antlions gain an additional second of life per kill
+	// Tooltip: Antlions gain {0} additional {^0:second;seconds} of life per kill
 }
 
 ThermalConversion: {
 	// Name: Thermal Conversion
-	// Tooltip: +0.5 seconds life gained per kill per level
+	// Tooltip: +{0} seconds life gained per kill per level
 }
 
 FrostbiteMandibles: {
 	// Name: Frostbite Mandibles
-	// Tooltip: Increased cold damage and a 10% chance to inflict Chilled on hit
+	// Tooltip: Increased cold damage and a {0}% chance to inflict Chilled on hit
 }
 
 AggressiveChill: {
